Conflicts:
	palladian-commons/pom.xml
	palladian-core/pom.xml
	palladian-experimental/pom.xml
	palladian-external-code/pom.xml
	palladian-external-liblinear/pom.xml
	palladian-persistence/pom.xml
	palladian-processing-api/pom.xml
	palladian-retrieval/pom.xml
	pom.xml

+++ palladian-retrieval/src/main/java/ws/palladian/retrieval/resources/WebVideo.java
@@ -15,7 +15,7 @@ public interface WebVideo extends WebContent {
     /**
      * @return The runtime of this video seconds, or <code>null</code> if no runtime was specified.
      */
-    Long getDuration();
+    Integer getDuration();
 
     /**
      * @return The number of views of this video, or <code>null</code> in case no view count was specified.
